chore: Put shared routes inside shared-bin
A lot of the routes are exactly the same for both binaries.
There is no reason for such duplication.
