+++ crates/maker/src/main.rs
@@ -197,15 +197,15 @@ async fn main() -> Result<()> {
                 routes::put_offer_params,
                 routes::put_offer_params_for_symbol,
                 routes::post_cfd_action,
-                routes::get_health_check,
                 routes::get_cfds,
-                routes::get_metrics,
                 routes::put_sync_wallet,
-                routes::get_version,
-                routes::change_password,
-                routes::post_login,
-                routes::logout,
-                routes::is_authenticated,
+                shared_bin::routes::get_health_check,
+                shared_bin::routes::get_metrics,
+                shared_bin::routes::get_version,
+                shared_bin::routes::change_password,
+                shared_bin::routes::logout,
+                shared_bin::routes::is_authenticated,
+                shared_bin::routes::post_login,
             ],
         )
         .register("/api", default_catchers())
