+++ crates/shared-bin/Cargo.toml
@@ -17,8 +17,10 @@ model = { path = "../model" }
 opentelemetry = { version = "0.18.0", features = ["rt-tokio"] }
 opentelemetry-otlp = "0.11.0"
 ping-pong = { path = "../xtra-libp2p-ping", package = "xtra-libp2p-ping" }
+prometheus = { version = "0.13", default-features = false }
 quiet-spans = { path = "../quiet-spans" }
 rocket = { version = "0.5.0-rc.2", features = ["json"] }
+rocket-cookie-auth = { path = "../rocket-cookie-auth" }
 serde = { version = "1", features = ["derive"] }
 time = "0.3.15"
 tracing = { version = "0.1" }
