+++ crates/taker/Cargo.toml
@@ -13,7 +13,6 @@ http-api-problem = { version = "0.55.0", features = ["rocket"] }
 itertools = "0.10"
 libp2p-core = { version = "0.33", default-features = false }
 model = { path = "../model" }
-prometheus = { version = "0.13", default-features = false }
 rocket = { version = "0.5.0-rc.2", features = ["json", "uuid"] }
 rocket-cookie-auth = { path = "../rocket-cookie-auth" }
 rocket-download-response = "0.5.2"
