+++ crates/taker/src/lib.rs
@@ -499,17 +499,17 @@ pub async fn run(opts: Opts) -> Result<()> {
             rocket::routes![
                 routes::feed,
                 routes::post_order_request,
-                routes::get_health_check,
                 routes::post_cfd_action,
                 routes::post_withdraw_request,
-                routes::get_metrics,
                 routes::put_sync_wallet,
-                routes::get_version,
-                routes::change_password,
-                routes::post_login,
-                routes::logout,
-                routes::is_authenticated,
                 routes::get_seed_backup,
+                shared_bin::routes::get_health_check,
+                shared_bin::routes::get_metrics,
+                shared_bin::routes::get_version,
+                shared_bin::routes::change_password,
+                shared_bin::routes::post_login,
+                shared_bin::routes::logout,
+                shared_bin::routes::is_authenticated,
             ],
         )
         .register("/api", default_catchers())
